# week7/week7-bepp-act2/backend-auth/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Job Management API",
    "version": "1.0.0",
    "description": "API for managing users and jobs, with authentication"
  },
  "servers": [
    {
      "url": "http://localhost:4000/api",
      "description": "Local Development Server"
    }
  ],
  "paths": {
    "/users/login": {
      "post": {
        "summary": "Login User",
        "operationId": "loginUser",
        "tags": [
          "User"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "user@example.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "password123"
                  }
                },
                "required": [
                  "email",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User logged in successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "email": {
                      "type": "string"
                    },
                    "token": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid credentials"
          }
        }
      }
    },
    "/users/signup": {
      "post": {
        "summary": "Signup User",
        "operationId": "signupUser",
        "tags": [
          "User"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  },
                  "phone_number": {
                    "type": "string"
                  },
                  "gender": {
                    "type": "string"
                  },
                  "date_of_birth": {
                    "type": "string",
                    "format": "date"
                  },
                  "membership_status": {
                    "type": "string"
                  }
                },
                "required": [
                  "name",
                  "email",
                  "password",
                  "phone_number",
                  "gender",
                  "date_of_birth",
                  "membership_status"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "email": {
                      "type": "string"
                    },
                    "token": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid user data"
          }
        }
      }
    },
    "/jobs": {
      "get": {
        "summary": "Get All Jobs",
        "operationId": "getAllJobs",
        "tags": [
          "Jobs"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of all jobs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Job"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "post": {
        "summary": "Create a Job",
        "operationId": "createJob",
        "tags": [
          "Jobs"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Job"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Job created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "400": {
            "description": "Invalid job data"
          }
        }
      }
    },
    "/jobs/{jobId}": {
      "get": {
        "summary": "Get Job by ID",
        "operationId": "getJobById",
        "tags": [
          "Jobs"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Job details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "404": {
            "description": "Job not found"
          }
        }
      },
      "put": {
        "summary": "Update Job by ID",
        "operationId": "updateJob",
        "tags": [
          "Jobs"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Job"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Job updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "404": {
            "description": "Job not found"
          }
        }
      },
      "delete": {
        "summary": "Delete Job by ID",
        "operationId": "deleteJob",
        "tags": [
          "Jobs"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Job deleted successfully"
          },
          "404": {
            "description": "Job not found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Job": {
        "type": "object",
        "required": [
          "title",
          "type",
          "description",
          "company"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "The job's unique ID"
          },
          "title": {
            "type": "string",
            "description": "The job's title"
          },
          "type": {
            "type": "string",
            "description": "The type of job (e.g., full-time, part-time)"
          },
          "description": {
            "type": "string",
            "description": "The description of the job"
          },
          "company": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string",
                "description": "The company's name"
              },
              "contactEmail": {
                "type": "string",
                "description": "The company's contact email"
              },
              "contactPhone": {
                "type": "string",
                "description": "The company's contact phone"
              }
            },
            "required": [
              "name",
              "contactEmail",
              "contactPhone"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Job's creation date"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Job's last update date"
          }
        }
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}